argument check in separate file
